+++ text/appetizers/README.md
@@ -2,4 +2,9 @@
 
 This section of the cookbook will focus on Appetizers, small runtime pallets that teach you the basics of writing pallets with a little hand-holding. If you are brand new to Substrate, you should follow through these appetizers in order. If you've already got the basics of pallet development down, you may skip ahead to the entrees which may be read in any order.
 
-<!-- TODO contents of this chapter? -->
+This section covers:
+
+* Dispatchable Calls
+* Events
+* Storage Values
+* Errors
